requirements.txt
MarkupSafe==2.1.1
python-dotenv==0.20.0
Werkzeug==2.2.2
